Backend/requirements.txt
flask_cors
mysql-connector-python